+++ cmd/abapEnvironmentAssembleConfirm_test.go
@@ -9,6 +9,23 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func TestPolling(t *testing.T) {
+	t.Run("Run polling", func(t *testing.T) {
+		var repo abaputils.Repository
+		b := testSetup(&abapbuild.ClMock{}, "ABIFNLDCSQPOVMXK4DNPBDRW2M")
+		var buildsWithRepo []buildWithRepository
+		bWR := buildWithRepository{
+			build: b,
+			repo:  repo,
+		}
+		buildsWithRepo = append(buildsWithRepo, bWR)
+		timeout := time.Duration(600 * time.Second)
+		pollInterval := time.Duration(1 * time.Second)
+		err := polling(buildsWithRepo, timeout, pollInterval)
+		assert.NoError(t, err)
+		assert.Equal(t, abapbuild.Finished, buildsWithRepo[0].build.RunState)
+	})
+}
 func TestStartingConfirm(t *testing.T) {
 	t.Run("Run starting", func(t *testing.T) {
 		client := &abapbuild.ClMock{
